requirements.txt
pandas
tensorflow
tda-api
coinbase
logging
pytest
configparser
backtrader
quantlib
ta-lib